requirements.txt
asgiref==3.9.1
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==5.2.6
django-cloudinary-storage==0.3.0
django-filter==25.2
django-vite==3.1.0
djangorestframework==3.16.1
idna==3.11
Markdown==3.9
pillow==12.0.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
